auth and users module separated, and isSessionActive function created
